Add `repl` filter and split main file into smaller modules.
