+++ jaq/src/filter.rs
@@ -1,0 +1,238 @@
+//! Filter parsing, compilation, and execution.
+use crate::{read, repl, Cli, Error};
+use core::fmt::{self, Display, Formatter};
+use jaq_core::{compile, load, Ctx, Native, RcIter, ValT};
+use jaq_json::Val;
+use std::io;
+use std::path::PathBuf;
+
+pub type Filter = jaq_core::Filter<Native<Val>>;
+
+pub fn parse_compile(
+    path: &PathBuf,
+    code: &str,
+    vars: &[String],
+    paths: &[PathBuf],
+) -> Result<(Vec<Val>, Filter), Vec<FileReports>> {
+    use compile::Compiler;
+    use load::{import, Arena, File, Loader};
+
+    let default = ["~/.jq", "$ORIGIN/../lib/jq", "$ORIGIN/../lib"].map(|x| x.into());
+    let paths = if paths.is_empty() { &default } else { paths };
+
+    let vars: Vec<_> = vars.iter().map(|v| format!("${v}")).collect();
+    let arena = Arena::default();
+    let defs = jaq_std::defs().chain(jaq_json::defs());
+    let loader = Loader::new(defs).with_std_read(paths);
+    //let loader = Loader::new([]).with_std_read(paths);
+    let path = path.into();
+    let modules = loader
+        .load(&arena, File { path, code })
+        .map_err(load_errors)?;
+
+    let mut vals = Vec::new();
+    import(&modules, |p| {
+        let path = p.find(paths, "json")?;
+        vals.push(read::json_array(path).map_err(|e| e.to_string())?);
+        Ok(())
+    })
+    .map_err(load_errors)?;
+
+    let funs = jaq_std::funs().chain(jaq_json::funs()).chain([repl::fun()]);
+    let compiler = Compiler::default()
+        .with_funs(funs)
+        .with_global_vars(vars.iter().map(|v| &**v));
+    let filter = compiler.compile(modules).map_err(compile_errors)?;
+    Ok((vals, filter))
+}
+
+/// Run a filter with given input values and run `f` for every value output.
+///
+/// This function cannot return an `Iterator` because it creates an `RcIter`.
+/// This is most unfortunate. We should think about how to simplify this ...
+pub(crate) fn run(
+    cli: &Cli,
+    filter: &Filter,
+    vars: Vec<Val>,
+    iter: impl Iterator<Item = io::Result<Val>>,
+    mut f: impl FnMut(Val) -> io::Result<()>,
+) -> Result<Option<bool>, Error> {
+    let mut last = None;
+    let iter = iter.map(|r| r.map_err(|e| e.to_string()));
+
+    let iter = Box::new(iter) as Box<dyn Iterator<Item = _>>;
+    let null = Box::new(core::iter::once(Ok(Val::Null))) as Box<dyn Iterator<Item = _>>;
+
+    let iter = RcIter::new(iter);
+    let null = RcIter::new(null);
+
+    let ctx = Ctx::new(vars, &iter);
+
+    for item in if cli.null_input { &null } else { &iter } {
+        let input = item.map_err(Error::Parse)?;
+        //println!("Got {:?}", input);
+        for output in filter.run((ctx.clone(), input)) {
+            let output = output.map_err(Error::Jaq)?;
+            last = Some(output.as_bool());
+            f(output)?;
+        }
+    }
+    Ok(last)
+}
+
+#[derive(Debug)]
+pub struct FileReports(load::File<String, PathBuf>, Vec<Report>);
+
+impl Display for FileReports {
+    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+        let Self(file, reports) = self;
+        let idx = codesnake::LineIndex::new(&file.code);
+        reports.iter().try_for_each(|e| {
+            writeln!(f, "Error: {}", e.message)?;
+            let block = e.to_block(&idx);
+            writeln!(f, "{}[{}]", block.prologue(), file.path.display())?;
+            writeln!(f, "{}{}", block, block.epilogue())
+        })
+    }
+}
+
+fn load_errors(errs: load::Errors<&str, PathBuf>) -> Vec<FileReports> {
+    use load::Error;
+
+    let errs = errs.into_iter().map(|(file, err)| {
+        let code = file.code;
+        let err = match err {
+            Error::Io(errs) => errs.into_iter().map(|e| report_io(code, e)).collect(),
+            Error::Lex(errs) => errs.into_iter().map(|e| report_lex(code, e)).collect(),
+            Error::Parse(errs) => errs.into_iter().map(|e| report_parse(code, e)).collect(),
+        };
+        FileReports(file.map_code(|s| s.into()), err)
+    });
+    errs.collect()
+}
+
+fn compile_errors(errs: compile::Errors<&str, PathBuf>) -> Vec<FileReports> {
+    let errs = errs.into_iter().map(|(file, errs)| {
+        let code = file.code;
+        let errs = errs.into_iter().map(|e| report_compile(code, e)).collect();
+        FileReports(file.map_code(|s| s.into()), errs)
+    });
+    errs.collect()
+}
+
+type StringColors = Vec<(String, Option<Color>)>;
+
+#[derive(Debug)]
+struct Report {
+    message: String,
+    labels: Vec<(core::ops::Range<usize>, StringColors, Color)>,
+}
+
+#[derive(Clone, Debug)]
+enum Color {
+    Yellow,
+    Red,
+}
+
+impl Color {
+    fn apply(&self, d: impl Display) -> String {
+        use yansi::{Color, Paint};
+        let color = match self {
+            Self::Yellow => Color::Yellow,
+            Self::Red => Color::Red,
+        };
+        d.fg(color).to_string()
+    }
+}
+
+fn report_io(code: &str, (path, error): (&str, String)) -> Report {
+    let path_range = load::span(code, path);
+    Report {
+        message: format!("could not load file {}: {}", path, error),
+        labels: [(path_range, [(error, None)].into(), Color::Red)].into(),
+    }
+}
+
+fn report_lex(code: &str, (expected, found): load::lex::Error<&str>) -> Report {
+    // truncate found string to its first character
+    let found = &found[..found.char_indices().nth(1).map_or(found.len(), |(i, _)| i)];
+
+    let found_range = load::span(code, found);
+    let found = match found {
+        "" => [("unexpected end of input".to_string(), None)].into(),
+        c => [("unexpected character ", None), (c, Some(Color::Red))]
+            .map(|(s, c)| (s.into(), c))
+            .into(),
+    };
+    let label = (found_range, found, Color::Red);
+
+    let labels = match expected {
+        load::lex::Expect::Delim(open) => {
+            let text = [("unclosed delimiter ", None), (open, Some(Color::Yellow))]
+                .map(|(s, c)| (s.into(), c));
+            Vec::from([(load::span(code, open), text.into(), Color::Yellow), label])
+        }
+        _ => Vec::from([label]),
+    };
+
+    Report {
+        message: format!("expected {}", expected.as_str()),
+        labels,
+    }
+}
+
+fn report_parse(code: &str, (expected, found): load::parse::Error<&str>) -> Report {
+    let found_range = load::span(code, found);
+
+    let found = if found.is_empty() {
+        "unexpected end of input"
+    } else {
+        "unexpected token"
+    };
+    let found = [(found.to_string(), None)].into();
+
+    Report {
+        message: format!("expected {}", expected.as_str()),
+        labels: Vec::from([(found_range, found, Color::Red)]),
+    }
+}
+
+fn report_compile(code: &str, (found, undefined): compile::Error<&str>) -> Report {
+    use compile::Undefined::Filter;
+    let found_range = load::span(code, found);
+    let wnoa = |exp, got| format!("wrong number of arguments (expected {exp}, found {got})");
+    let message = match (found, undefined) {
+        ("reduce", Filter(arity)) => wnoa("2", arity),
+        ("foreach", Filter(arity)) => wnoa("2 or 3", arity),
+        (_, undefined) => format!("undefined {}", undefined.as_str()),
+    };
+    let found = [(message.clone(), None)].into();
+
+    Report {
+        message,
+        labels: Vec::from([(found_range, found, Color::Red)]),
+    }
+}
+
+type CodeBlock = codesnake::Block<codesnake::CodeWidth<String>, String>;
+
+impl Report {
+    fn to_block(&self, idx: &codesnake::LineIndex) -> CodeBlock {
+        use codesnake::{Block, CodeWidth, Label};
+        let color_maybe = |(text, color): (_, Option<Color>)| match color {
+            None => text,
+            Some(color) => color.apply(text).to_string(),
+        };
+        let labels = self.labels.iter().cloned().map(|(range, text, color)| {
+            let text = text.into_iter().map(color_maybe).collect::<Vec<_>>();
+            Label::new(range)
+                .with_text(text.join(""))
+                .with_style(move |s| color.apply(s).to_string())
+        });
+        Block::new(idx, labels).unwrap().map_code(|c| {
+            let c = c.replace('\t', "    ");
+            let w = unicode_width::UnicodeWidthStr::width(&*c);
+            CodeWidth::new(c, core::cmp::max(w, 1))
+        })
+    }
+}
